[chore] 폴더 위치 및 약간 수정 1. 과제 파일들을 assignment 폴더로 이동 2. 클레릭 테스트 마나 수치를 현존 마나 수치로 수정 3. 법사 마나 오링 현상 mp 조정(6 -> 5)
